backend/requirements.txt
djangorestframework-simplejwt>=4.3.0,<5.0.0
pyjwt==2.8.0
djoser==2.1.0
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # PostgreSQL接続用
dj-database-url==2.1.0  # データベースURL解析用
gunicorn==21.2.0  # 本番環境用サーバー
whitenoise==6.6.0  # 静的ファイル配信
google-cloud-aiplatform==1.71.0  # Vertex AI SDK
vertexai==1.71.0  # Vertex AI Python SDK
